feat: main refactor
