+++ src/Renderer/Renderer.h
@@ -5,4 +5,5 @@
 class Renderer {
 public:
     virtual void renderFrame(const uint8_t *frame) = 0;
+    virtual bool shouldStopRendering() = 0;
 };
